requirements.txt
Adafruit-Python-DHT==1.1.2
paho-mqtt==1.2
configparser==3.5.0
